+++ Nos/Models/CoreData/Event+CoreDataClass.swift
@@ -903,7 +903,11 @@ public class Event: NosManagedObject {
         seenOnRelays.compactMap { $0.addressURL?.absoluteString }
     }
     
-    class func attributedContent(noteID: String?, context: NSManagedObjectContext) async -> AttributedString {
+    class func attributedContent(
+        noteID: String?,
+        noteParser: NoteParser = NoteParser(),
+        context: NSManagedObjectContext
+    ) async -> AttributedString {
         guard let noteID else {
             return AttributedString()
         }
@@ -915,7 +919,7 @@ public class Event: NosManagedObject {
             }
             try? context.saveIfNeeded()
             let tags = note.allTags as? [[String]] ?? []
-            return NoteParser().parse(
+            return noteParser.parse(
                 content: content,
                 tags: tags,
                 context: context
@@ -957,7 +961,8 @@ public class Event: NosManagedObject {
     /// - Returns: A tuple where the first object is the note content formatted for display, and the second is a list
     ///     of HTTP links found in the note's context.  
     @MainActor class func attributedContentAndURLs(
-        note: Event, 
+        note: Event,
+        noteParser: NoteParser = NoteParser(),
         context: NSManagedObjectContext
     ) async -> (AttributedString, [URL])? {
         guard let content = note.content else {
@@ -966,7 +971,7 @@ public class Event: NosManagedObject {
         let tags = note.allTags as? [[String]] ?? []
         
         return await context.perform {
-            NoteParser().parse(content: content, tags: tags, context: context)
+            noteParser.parse(content: content, tags: tags, context: context)
         }
     }
