+++ NosTests/NoteParserTests.swift
@@ -2,12 +2,10 @@ import CoreData
 import XCTest
 import Dependencies
 
-// swiftlint:disable file_length
-
-final class NoteNoteParserTests: CoreDataTestCase {
+final class NoteParserTests: CoreDataTestCase {
 
     // swiftlint:disable:next implicitly_unwrapped_optional
-    private var sut: NoteParser!
+    var sut: NoteParser!
 
     override func setUp() async throws {
         sut = NoteParser()
@@ -85,82 +83,6 @@ final class NoteNoteParserTests: CoreDataTestCase {
         XCTAssertEqual(links[safe: 0]?.key, nip05)
         XCTAssertEqual(links[safe: 0]?.value, URL(string: webLink))
     }
-
-    func testMentionPrecededByAt() throws {
-        // Arrange
-        let name = "nos"
-        let npub = "npub1pu3vqm4vzqpxsnhuc684dp2qaq6z69sf65yte4p39spcucv5lzmqswtfch"
-        let hex = "0f22c06eac1002684efcc68f568540e8342d1609d508bcd4312c038e6194f8b6"
-
-        // The @ symbol before nostr: should not break the parsing
-        let content = "Ping @nostr:\(npub)"
-        let expected = "Ping @\(name)"
-
-        let tags = [
-            ["p", hex],
-            ["p", "8c430bdaadc1a202e4dd11c86c82546bb108d755e374b7918181f533b94e312e"],
-            ["e", "a9788ca56a90bb5b856e89f16f5f3b0da93c28ea625e845c9925a41377152a13", "", "root"],
-            ["e", "3d9503a2d4ad024749b138c041e99934474e2822e2a1c697792dab5b24acc285", "", "reply"]
-        ]
-
-        // Act
-        let context = try XCTUnwrap(testContext)
-        let author = try Author.findOrCreate(by: hex, context: context)
-        author.displayName = name
-        try context.save()
-
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-
-        // Assert
-        let parsedContent = String(attributedContent.characters)
-        XCTAssertEqual(parsedContent, expected)
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(name)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
-
-    func testMentionToProfileWithURLInName() throws {
-        // Arrange
-        let name = "nos.social" // This should not break the parsing
-        let npub = "npub1pu3vqm4vzqpxsnhuc684dp2qaq6z69sf65yte4p39spcucv5lzmqswtfch"
-        let hex = "0f22c06eac1002684efcc68f568540e8342d1609d508bcd4312c038e6194f8b6"
-        
-        let content = "Yep. Something like that. I, of course could implement it " +
-            "in nostr:\(npub) but haven’t yet."
-        let expected = "Yep. Something like that. I, of course could implement it " +
-            "in @\(name) but haven’t yet."
-
-        let tags = [
-            ["p", hex],
-            ["p", "8c430bdaadc1a202e4dd11c86c82546bb108d755e374b7918181f533b94e312e"],
-            ["e", "a9788ca56a90bb5b856e89f16f5f3b0da93c28ea625e845c9925a41377152a13", "", "root"],
-            ["e", "3d9503a2d4ad024749b138c041e99934474e2822e2a1c697792dab5b24acc285", "", "reply"]
-        ]
-
-        // Act
-        let context = try XCTUnwrap(testContext)
-        let author = try Author.findOrCreate(by: hex, context: context)
-        author.displayName = name
-        try context.save()
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-
-        // Assert
-        let parsedContent = String(attributedContent.characters)
-        XCTAssertEqual(parsedContent, expected)
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(name)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
     
     /// Example taken from [NIP-27](https://github.com/nostr-protocol/nips/blob/master/27.md)
     func testMentionWithNPub() throws {
@@ -170,193 +92,15 @@ final class NoteNoteParserTests: CoreDataTestCase {
         let link = "nostr:\(npub)"
         let markdown = "hello [\(mention)](\(link))"
         let attributedString = try AttributedString(markdown: markdown)
-        let (content, tags) = sut.parse(attributedText: attributedString)
+        let (content, tags) = sut.parse(
+            attributedText: attributedString
+        )
         let expectedContent = "hello nostr:\(npub)"
         let expectedTags = [["p", hex]]
         XCTAssertEqual(content, expectedContent)
         XCTAssertEqual(tags, expectedTags)
     }
-
-    func testContentWithNIP08Mention() throws {
-        let name = "mattn"
-        let content = "hello #[0]"
-        let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
-        let expectedContent = "hello @\(name)"
-        let tags = [["p", hex]]
-        let context = try XCTUnwrap(testContext)
-        let author = try Author.findOrCreate(by: hex, context: context)
-        author.displayName = name
-        try context.save()
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let parsedContent = String(attributedContent.characters)
-        XCTAssertEqual(parsedContent, expectedContent)
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(name)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
-
-    func testContentWithNIP08MentionToUnknownAuthor() async throws {
-        let content = "hello #[0]"
-        let displayName = "npub1937vv..."
-        let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
-        let expectedContent = "hello @\(displayName)"
-        let tags = [["p", hex]]
-        let context = persistenceController.viewContext
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let parsedContent = String(attributedContent.characters)
-        XCTAssertEqual(parsedContent, expectedContent)
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(displayName)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
-
-    func testContentWithNIP08MentionAtBeginning() throws {
-        let content = "#[0]"
-        let displayName = "npub1937vv..."
-        let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
-        let expectedContent = "@\(displayName)"
-        let tags = [["p", hex]]
-        let context = try XCTUnwrap(testContext)
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let parsedContent = String(attributedContent.characters)
-        XCTAssertEqual(parsedContent, expectedContent)
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(displayName)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
-
-    func testContentWithNIP08MentionAfterNewline() throws {
-        let content = "Hello\n#[0]"
-        let displayName = "npub1937vv..."
-        let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
-        let expectedContent = "Hello\n@\(displayName)"
-        let tags = [["p", hex]]
-        let context = try XCTUnwrap(testContext)
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let parsedContent = String(attributedContent.characters)
-        XCTAssertEqual(parsedContent, expectedContent)
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(displayName)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
-
-    func testContentWithNIP08MentionInsideAWord() throws {
-        let content = "hello#[0]"
-        let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
-        let expectedContent = "hello#[0]"
-        let tags = [["p", hex]]
-        let context = try XCTUnwrap(testContext)
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let parsedContent = String(attributedContent.characters)
-        XCTAssertEqual(parsedContent, expectedContent)
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 0)
-    }
-
-    /// Example taken from [NIP-27](https://github.com/nostr-protocol/nips/blob/master/27.md)
-    func testContentWithNIP27Mention() throws {
-        let name = "mattn"
-        let content = "hello nostr:npub1937vv2nf06360qn9y8el6d8sevnndy7tuh5nzre4gj05xc32tnwqauhaj6"
-        let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
-        let tags = [["p", hex]]
-        let context = try XCTUnwrap(testContext)
-        let author = try Author.findOrCreate(by: hex, context: context)
-        author.displayName = name
-        try context.save()
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(name)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
-
-    /// Example taken from [NIP-27](https://github.com/nostr-protocol/nips/blob/master/27.md)
-    func testContentWithNIP27MentionToUnknownAuthor() throws {
-        let displayName = "npub1937vv..."
-        let content = "hello nostr:npub1937vv2nf06360qn9y8el6d8sevnndy7tuh5nzre4gj05xc32tnwqauhaj6"
-        let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
-        let tags = [["p", hex]]
-        let context = try XCTUnwrap(testContext)
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(displayName)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
-
-    /// Example taken from [NIP-27](https://github.com/nostr-protocol/nips/blob/master/27.md)
-    func testContentWithNIP27ProfileMention() throws {
-        let name = "mattn"
-        let content = "hello nostr:nprofile1qqszclxx9f5haga8sfjjrulaxncvkfekj097t6f3pu65f86rvg49ehqj6f9dh"
-        let hex = "2c7cc62a697ea3a7826521f3fd34f0cb273693cbe5e9310f35449f43622a5cdc"
-        let tags = [["p", hex]]
-        let context = try XCTUnwrap(testContext)
-        let author = try Author.findOrCreate(by: hex, context: context)
-        author.displayName = name
-        try context.save()
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(name)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
-
-    func testContentWithNIP27ProfileMentionWithADot() throws {
-        let name = "mattn"
-        let content = "Hello nostr:npub1pu3vqm4vzqpxsnhuc684dp2qaq6z69sf65yte4p39spcucv5lzmqswtfch.\n\nBye"
-        let hex = "0f22c06eac1002684efcc68f568540e8342d1609d508bcd4312c038e6194f8b6"
-        let tags = [["p", hex]]
-        let context = try XCTUnwrap(testContext)
-        let author = try Author.findOrCreate(by: hex, context: context)
-        author.displayName = name
-        try context.save()
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 1)
-        XCTAssertEqual(links.first?.key, "@\(name)")
-        XCTAssertEqual(links.first?.value, URL(string: "@\(hex)"))
-    }
-
+    
     func testContentWithMixedMentions() throws {
         let content = "hello nostr:npub1937vv2nf06360qn9y8el6d8sevnndy7tuh5nzre4gj05xc32tnwqauhaj6 and #[1]"
         let displayName1 = "npub1937vv..."
@@ -425,28 +169,6 @@ final class NoteNoteParserTests: CoreDataTestCase {
         XCTAssertEqual(links.count, 1)
         XCTAssertEqual(links[safe: 0]?.key, "🔗 Link to note")
         XCTAssertEqual(links[safe: 0]?.value, URL(string: "%\(hex)"))
-    }
-
-    func testContentWithUntaggedNIP27NoteAndTaggedNIP27Profile() throws {
-        let profileDisplayName = "@npub1pu3vq..."
-        let profile = "npub1pu3vqm4vzqpxsnhuc684dp2qaq6z69sf65yte4p39spcucv5lzmqswtfch"
-        let note = "note1h2mmqfjqle48j8ytmdar22v42g5y9n942aumyxatgtxpqj29pjjsjecraw"
-        let content = "Check this nostr:\(note) from nostr:\(profile)"
-        let profileHex = "0f22c06eac1002684efcc68f568540e8342d1609d508bcd4312c038e6194f8b6"
-        let noteHex = "bab7b02640fe6a791c8bdb7a352995522842ccb55779b21bab42cc1049450ca5"
-        let tags: [[String]] = [["p", profileHex]]
-        let context = try XCTUnwrap(testContext)
-        let (attributedContent, _) = sut.parse(
-            content: content,
-            tags: tags,
-            context: context
-        )
-        let links = attributedContent.links
-        XCTAssertEqual(links.count, 2)
-        XCTAssertEqual(links[safe: 0]?.key, "🔗 Link to note")
-        XCTAssertEqual(links[safe: 0]?.value, URL(string: "%\(noteHex)"))
-        XCTAssertEqual(links[safe: 1]?.key, "\(profileDisplayName)")
-        XCTAssertEqual(links[safe: 1]?.value, URL(string: "@\(profileHex)"))
     }
     
     func testContentWithUntaggedProfile() throws {
@@ -548,16 +270,5 @@ final class NoteNoteParserTests: CoreDataTestCase {
 
         let links = attributedContent.links
         XCTAssertEqual(links.count, 0)
-    }
-}
-
-fileprivate extension AttributedString {
-    var links: [(key: String, value: URL)] {
-        runs.compactMap {
-            guard let link = $0.link else {
-                return nil
-            }
-            return (key: String(self[$0.range].characters), value: link)
-        }
     }
 }
